REM-55 feat: добавление тестов для логики закрытия, рендеринга и сохранения компонента Settings в нормальном режиме
