--- frontend/src/components/settings/__tests__/Settings.NormalMode.test.tsx
@@ -1,347 +1,0 @@
-// Tests for Settings component - Normal Mode
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { render, screen, fireEvent, waitFor } from '@testing-library/react';
-import React from 'react';
-import { Settings, SettingsProps } from '../Settings';
-import { LoadConfig, SaveAllSettings } from '../../../../wailsjs/go/main/App';
-import { ConnectionConfig } from '../../../App';
-
-// --- Mocks ---
-
-// Mock child components
-vi.mock('../ConnectionTab', () => ({
-    ConnectionTab: vi.fn(({ settings, onSettingsChange, onConnectionTest, errors }) => (
-        <div data-testid="connection-tab-mock">
-            Connection Tab Mock
-            <input
-                data-testid="connection-host-input"
-                value={settings.host}
-                onChange={(e) => onSettingsChange({ host: e.target.value })}
-            />
-            <button data-testid="connection-test-button-success" onClick={() => onConnectionTest(true)}>Test Success</button>
-            <button data-testid="connection-test-button-fail" onClick={() => onConnectionTest(false, 'Custom error message')}>Test Fail</button>
-            {errors?.host && <span data-testid="connection-host-error">{errors.host}</span>}
-        </div>
-    )),
-}));
-
-vi.mock('../LimitsTab', () => ({
-    LimitsTab: vi.fn(({ settings, onSettingsChange, errors }) => (
-        <div data-testid="limits-tab-mock">
-            Limits Tab Mock
-            <input
-                data-testid="limits-ratio-input"
-                type="number"
-                value={settings.maxUploadRatio}
-                onChange={(e) => onSettingsChange({ maxUploadRatio: Number(e.target.value) })}
-            />
-            {errors?.maxUploadRatio && <span data-testid="limits-ratio-error">{errors.maxUploadRatio}</span>}
-        </div>
-    )),
-}));
-
-// Mock PathsTab and its ref methods
-const mockGetPathChanges = vi.fn(() => ({ pathsToAdd: [], pathsToRemove: [], defaultPath: '' }));
-const mockResetChanges = vi.fn();
-
-interface PathsTabProps {
-    onPathsChanged: (hasChanges: boolean) => void;
-}
-
-vi.mock('../PathsTab', () => ({
-    PathsTab: vi.fn(React.forwardRef(({ onPathsChanged }: PathsTabProps, ref) => {
-        React.useImperativeHandle(ref, () => ({
-            getPathChanges: mockGetPathChanges,
-            resetChanges: mockResetChanges,
-        }));
-        return (
-            <div data-testid="paths-tab-mock">
-                Paths Tab Mock
-                <button data-testid="paths-change-button" onClick={() => onPathsChanged(true)}>Change Path</button>
-            </div>
-        );
-    })),
-}));
-
-vi.mock('../../LanguageSelector', () => ({
-    LanguageSelector: vi.fn(() => <div data-testid="language-selector-mock">Language Selector Mock</div>),
-}));
-
-vi.mock('../../StatusMessage', () => ({
-    default: vi.fn(({ status, message, 'data-testid': dataTestId }) => (
-        <div data-testid={dataTestId || 'status-message-mock'} data-status={status}>
-            {message}
-        </div>
-    )),
-}));
-
-vi.mock('../../LoadingSpinner', () => ({
-    LoadingSpinner: vi.fn(() => <div data-testid="loading-spinner-mock">Loading...</div>),
-}));
-
-// Mock useLocalization - relying on global mock
-
-// Mock useSettingsSaver - needed for isSaving state
-// We'll mock it specifically in the test that needs it.
-
-// Default props for Normal Mode tests
-const defaultProps: SettingsProps = {
-    onSave: vi.fn().mockResolvedValue(true), // Not directly used in normal mode save flow
-    onClose: vi.fn(),
-    isFirstStart: false, // Explicitly set for this test suite
-};
-
-// Helper function to render the component
-const renderSettings = (props: Partial<SettingsProps> = {}) => {
-    // Ensure isFirstStart is always false for this suite
-    return render(<Settings {...defaultProps} {...props} isFirstStart={false} />);
-};
-
-// Mock config for loading
-const mockConfig: ConnectionConfig = {
-    host: 'test-host', port: 1234, username: 'user', password: 'pw',
-    maxUploadRatio: 2, slowSpeedLimit: 100, slowSpeedUnit: 'MiB/s'
-};
-
-// --- Tests ---
-
-describe('Settings Component - Normal Mode', () => {
-    // Reset mocks before each test
-    beforeEach(() => {
-        vi.clearAllMocks();
-        // Provide default mock implementations for Wails functions
-        (LoadConfig as unknown as ReturnType<typeof vi.fn>).mockResolvedValue(mockConfig); // Default: load mock config
-        (SaveAllSettings as unknown as ReturnType<typeof vi.fn>).mockResolvedValue(undefined); // Default: save succeeds
-        (defaultProps.onSave as unknown as ReturnType<typeof vi.fn>).mockClear().mockResolvedValue(true);
-        (defaultProps.onClose as unknown as ReturnType<typeof vi.fn>).mockClear();
-        mockGetPathChanges.mockClear().mockReturnValue({ pathsToAdd: [], pathsToRemove: [], defaultPath: '' });
-        mockResetChanges.mockClear();
-    });
-
-    afterEach(() => {
-        // Clean up any potential module-level mocks if set within tests
-        vi.unmock('../hooks/useSettingsSaver');
-    });
-
-    // Test initial rendering in loading state
-    it('renders loading spinner initially', async () => {
-        // Mock LoadConfig to simulate loading delay
-        (LoadConfig as unknown as ReturnType<typeof vi.fn>).mockImplementation(
-            () => new Promise(resolve => setTimeout(() => resolve(mockConfig), 50))
-        );
-        renderSettings();
-
-        // Expect loading spinner to be present initially
-        expect(screen.getByTestId('settings-loading')).toBeInTheDocument();
-        expect(screen.getByTestId('loading-spinner-mock')).toBeInTheDocument();
-
-        // Wait for loading to complete
-        await waitFor(() => {
-            expect(screen.queryByTestId('settings-loading')).not.toBeInTheDocument();
-        });
-    });
-
-    // Test rendering in normal mode (after loading)
-    it('renders correctly after loading', async () => {
-        renderSettings();
-        await waitFor(() => expect(screen.queryByTestId('settings-loading')).not.toBeInTheDocument());
-
-        // Check title
-        expect(screen.getByTestId('settings-title')).toHaveTextContent('settings.title');
-
-        // Check no first start message or language selector
-        expect(screen.queryByTestId('settings-first-start-message')).not.toBeInTheDocument();
-        expect(screen.queryByTestId('language-selector-mock')).not.toBeInTheDocument();
-
-        // Check tabs - all should be visible
-        expect(screen.getByTestId('settings-tab-connection')).toBeInTheDocument();
-        expect(screen.getByTestId('settings-tab-limits')).toBeInTheDocument();
-        expect(screen.getByTestId('settings-tab-paths')).toBeInTheDocument();
-
-        // Check buttons - Save and Cancel should be visible
-        expect(screen.getByTestId('settings-save-button')).toBeInTheDocument();
-        expect(screen.getByTestId('settings-cancel-button')).toBeInTheDocument();
-
-        // Check if ConnectionTab mock is rendered by default
-        expect(screen.getByTestId('connection-tab-mock')).toBeInTheDocument();
-
-        // Save button should be enabled initially in normal mode
-        expect(screen.getByTestId('settings-save-button')).not.toBeDisabled();
-    });
-
-    // Test tab switching
-    it('switches between tabs correctly', async () => {
-        renderSettings();
-        await waitFor(() => expect(screen.queryByTestId('settings-loading')).not.toBeInTheDocument());
-
-        // Initially, Connection tab is active
-        expect(screen.getByTestId('connection-tab-mock')).toBeInTheDocument();
-        expect(screen.queryByTestId('limits-tab-mock')).not.toBeInTheDocument();
-        expect(screen.queryByTestId('paths-tab-mock')).not.toBeInTheDocument();
-
-        // Click Limits tab
-        fireEvent.click(screen.getByTestId('settings-tab-limits'));
-        await waitFor(() => {
-            expect(screen.queryByTestId('connection-tab-mock')).not.toBeInTheDocument();
-            expect(screen.getByTestId('limits-tab-mock')).toBeInTheDocument();
-            expect(screen.queryByTestId('paths-tab-mock')).not.toBeInTheDocument();
-        });
-
-        // Click Paths tab
-        fireEvent.click(screen.getByTestId('settings-tab-paths'));
-        await waitFor(() => {
-            expect(screen.queryByTestId('connection-tab-mock')).not.toBeInTheDocument();
-            expect(screen.queryByTestId('limits-tab-mock')).not.toBeInTheDocument();
-            expect(screen.getByTestId('paths-tab-mock')).toBeInTheDocument();
-        });
-
-        // Click Connection tab again
-        fireEvent.click(screen.getByTestId('settings-tab-connection'));
-        await waitFor(() => {
-            expect(screen.getByTestId('connection-tab-mock')).toBeInTheDocument();
-            expect(screen.queryByTestId('limits-tab-mock')).not.toBeInTheDocument();
-            expect(screen.queryByTestId('paths-tab-mock')).not.toBeInTheDocument();
-        });
-    });
-
-    // Test settings changes within tabs
-    it('updates settings state when changed in tabs', async () => {
-        renderSettings();
-        await waitFor(() => expect(screen.queryByTestId('settings-loading')).not.toBeInTheDocument());
-
-        // Change host in ConnectionTab
-        const hostInput = screen.getByTestId('connection-host-input');
-        fireEvent.change(hostInput, { target: { value: 'new-host' } });
-        await waitFor(() => expect(hostInput).toHaveValue('new-host'));
-
-        // Switch to Limits tab
-        fireEvent.click(screen.getByTestId('settings-tab-limits'));
-        await waitFor(() => expect(screen.getByTestId('limits-tab-mock')).toBeInTheDocument());
-
-        // Change ratio in LimitsTab
-        const ratioInput = screen.getByTestId('limits-ratio-input');
-        fireEvent.change(ratioInput, { target: { value: '10' } });
-        await waitFor(() => expect(ratioInput).toHaveValue(10));
-
-        // Saving logic is tested in useSettingsSaver.test.ts
-    });
-
-    // Test Cancel button logic
-    it('calls onClose and resets path changes when Cancel is clicked', async () => {
-        renderSettings();
-        await waitFor(() => expect(screen.queryByTestId('settings-loading')).not.toBeInTheDocument());
-
-        expect(mockResetChanges).not.toHaveBeenCalled();
-        const cancelButton = screen.getByTestId('settings-cancel-button');
-        fireEvent.click(cancelButton);
-
-        expect(defaultProps.onClose).toHaveBeenCalledTimes(1);
-        expect(mockResetChanges).toHaveBeenCalledTimes(1);
-    });
-
-    // Test closing via onOpenChange (e.g., Escape key)
-    it('calls onClose and resets path changes when dialog requests close', async () => {
-        renderSettings();
-        await waitFor(() => expect(screen.queryByTestId('settings-loading')).not.toBeInTheDocument());
-
-        const dialogContent = screen.getByRole('dialog');
-        fireEvent.keyDown(dialogContent, { key: 'Escape', code: 'Escape' });
-
-        expect(defaultProps.onClose).toHaveBeenCalledTimes(1);
-        expect(mockResetChanges).toHaveBeenCalledTimes(1);
-    });
-
-    // Test Save button disabled state when saving
-    it('disables Save button when saving', async () => {
-        // Mock useSettingsSaver to control isSaving state
-        vi.doMock('../hooks/useSettingsSaver', () => ({
-            useSettingsSaver: vi.fn(() => ({
-                isSaving: true, // Force isSaving to true
-                handleSave: vi.fn(),
-            })),
-        }));
-
-        // Dynamically import Settings again to use the mocked hook
-        const { Settings: SettingsWithMockedSaver } = await import('../Settings');
-
-        render(<SettingsWithMockedSaver {...defaultProps} isFirstStart={false} />);
-        await waitFor(() => expect(screen.queryByTestId('settings-loading')).not.toBeInTheDocument());
-
-        // Check if the button is disabled due to isSaving=true
-        await waitFor(() => {
-            expect(screen.getByTestId('settings-save-button')).toBeDisabled();
-        });
-
-        // Clean up the mock
-        vi.unmock('../hooks/useSettingsSaver');
-    });
-
-    // Test dialog closing prevention (onPointerDownOutside)
-    it('prevents closing and does not call onClose when clicking outside if saving', async () => {
-        // Mock useSettingsSaver to control isSaving state
-        vi.doMock('../hooks/useSettingsSaver', () => ({
-            useSettingsSaver: vi.fn(() => ({
-                isSaving: true, // Force isSaving to true
-                handleSave: vi.fn(),
-            })),
-        }));
-
-        const { Settings: SettingsWithMockedSaver } = await import('../Settings');
-        render(<SettingsWithMockedSaver {...defaultProps} isFirstStart={false} />);
-        await waitFor(() => expect(screen.queryByTestId('settings-loading')).not.toBeInTheDocument());
-
-        // Simulate click outside - Radix Dialog calls onPointerDownOutside on the Content
-        // We need to trigger the event handler passed to DialogContent
-        // This requires finding the DialogContent and simulating the event dispatch
-        const dialogContent = screen.getByRole('dialog');
-
-        // Create a mock event that has preventDefault
-        const mockEvent = { preventDefault: vi.fn() };
-
-        // Find the component instance or props to call the handler
-        // This is difficult with testing-library alone. We assume the handler is attached.
-        // Let's simulate the *effect* - preventDefault should be called, onClose should not.
-
-        // We can't directly call the prop function easily here.
-        // Instead, we test that onClose is NOT called after a simulated external interaction
-        // (like clicking the body, although Radix handles this internally).
-        // Let's assume Radix calls the handler we provide.
-
-        // A better test might involve checking if preventDefault was called on the event,
-        // but that requires deeper integration testing or component internals access.
-
-        // For now, let's verify onClose is not called after a delay, assuming the outside click happened.
-        fireEvent.pointerDown(document.body); // Simulate a click somewhere outside
-
-        // Wait a bit to ensure async operations (if any) complete
-        await new Promise(resolve => setTimeout(resolve, 50));
-
-        expect(defaultProps.onClose).not.toHaveBeenCalled();
-
-        // Clean up the mock
-        vi.unmock('../hooks/useSettingsSaver');
-    });
-
-    it('allows closing and calls onClose when clicking outside if not saving', async () => {
-        // Use default useSettingsSaver mock (isSaving = false)
-        renderSettings();
-        await waitFor(() => expect(screen.queryByTestId('settings-loading')).not.toBeInTheDocument());
-
-        // Simulate click outside
-        fireEvent.pointerDown(document.body);
-
-        // In a real scenario, Radix would trigger onOpenChange(false) which calls our handleClose.
-        // Since we can't easily simulate the Radix internal behavior perfectly,
-        // we'll assume the click outside *would* lead to handleClose being called if not prevented.
-        // The previous test verifies prevention when isSaving=true.
-        // This test implicitly verifies non-prevention by checking if onClose *would* be called
-        // if Radix triggered it. Let's simulate the Escape key again as a proxy for allowed closing.
-
-        const dialogContent = screen.getByRole('dialog');
-        fireEvent.keyDown(dialogContent, { key: 'Escape', code: 'Escape' });
-
-        expect(defaultProps.onClose).toHaveBeenCalledTimes(1);
-        expect(mockResetChanges).toHaveBeenCalledTimes(1); // Ensure reset is also called
-    });
-
-});
